// multi/src/app/model/motion-events.ts
/**********************************************************************************
* File Name   :   MotionEvents.ts
* Description :   MotionEvents
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class MotionEvents {
    keyframeTime: string;
    startTime: string;
    endTime: string;
    jumpTime: number;
}