// multi/src/app/model/recording-events.ts
/**********************************************************************************
* File Name   :   RecordingEvents.ts
* Description :   RecordingEvents
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class RecordingEvents {
    startTime: string;
    endTime: string;
}