// multi/src/app/model/player-thumbnail.ts
/**********************************************************************************
* File Name   :   PlayerThumbnail.ts
* Description :   PlayerThumbnail
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class PlayerThumbnail {
    lowResUrl: string;
    highResUrl: string;
    startTime: string;
}