// multi/src/app/model/playback-baseurl.ts
/**********************************************************************************
* File Name   :   PlaybackBaseurl.ts
* Description :   PlaybackBaseurl
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class PlaybackBaseurl {
    baseUrl: string;
    gatewayIP: string;
 }